etxweb-rest/src/main/java/eu/europa/ec/etrustex/web/rest/resolver/UserDetailsHandlerMethodArgumentResolver.java
package eu.europa.ec.etrustex.web.rest.resolver;

import eu.europa.ec.etrustex.web.security.SecurityUserDetails;
import lombok.extern.slf4j.Slf4j;
import org.springframework.core.MethodParameter;
import org.springframework.security.access.AccessDeniedException;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.stereotype.Component;
import org.springframework.web.bind.support.WebDataBinderFactory;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.method.support.HandlerMethodArgumentResolver;
import org.springframework.web.method.support.ModelAndViewContainer;

import javax.validation.constraints.NotNull;

@Component
@Slf4j
public class UserDetailsHandlerMethodArgumentResolver implements HandlerMethodArgumentResolver {
    @Override
    public boolean supportsParameter(@NotNull MethodParameter parameter) {
        return SecurityUserDetails.class.isAssignableFrom(parameter.getParameterType());
    }

    @Override
    public Object resolveArgument(@NotNull MethodParameter parameter, ModelAndViewContainer mavContainer, @NotNull NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
        log.debug("Authentication: " + authentication);

        try {
            return authentication.getPrincipal();
        } catch (Exception e) {
            throw new AccessDeniedException(String.format("No Authentication for context path %s. Is it configured in SecurityConfig antMatchers?", webRequest.getContextPath()), e);
        }
    }
}
